% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{cov.mat.check}
\alias{cov.mat.check}
\title{Check each covariance matrix in the gaussian mixture specification}
\usage{
cov.mat.check(model)
}
\arguments{
\item{model}{Hext model structure with an non-empty M5 slot}
}
\value{
None
}
\description{
Display the sum of log eigenvalues of each covariance matrix
}

