% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density_fun.r
\name{cov.wt.hext}
\alias{cov.wt.hext}
\title{Weighted covariance matrices, similar to cov.wt in hsmm}
\usage{
cov.wt.hext(x, wt = rep(1/nrow(x), nrow(x)), cor = FALSE, center = TRUE,
  method = c("unbiased", "ML"))
}
\arguments{
\item{x}{numeric matrix of size T x p, where T is number of obs, p the dimension of x}

\item{wt}{vector of weights for each obs}

\item{cor}{boolean whether correlation matrix to be returned}

\item{center}{centers to be used when computing covariances}

\item{method}{either "unbiased" or "ML}
}
\value{
list with components cov, center, n.obs, wt, cor
}
\description{
Weighted covariance matrices, similar to cov.wt in hsmm
}

