% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density_fun.r
\name{dbeta.hext}
\alias{dbeta.hext}
\title{Beta emission density for hidden state j in an HMM}
\usage{
dbeta.hext(x, j, model)
}
\arguments{
\item{x}{numeric vector}

\item{j}{hidden state index}

\item{model}{HMM model}
}
\value{
vector of proabilities, length equal the number of rows in x
}
\description{
Beta emission density for hidden state j in an HMM
}

