% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{density.ndim}
\alias{density.ndim}
\title{Compute the density of points}
\usage{
\method{density}{ndim}(dat)
}
\arguments{
\item{dat}{numeric matrix of size n x p, where n is the nubmer of points in p-dimensional space}
}
\value{
numeric value which is the density in p-dim ellipsoid
}
\description{
compute the density in an ellipse
}

