% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{dist.one}
\alias{dist.one}
\title{Computes the euclidean distances for n pairs of points}
\usage{
dist.one(tgt, pts)
}
\arguments{
\item{tgt}{target, a vector of length p}

\item{pts}{n x p matrix, where each row is a point in p-dimensional space}
}
\value{
vector of distances
}
\description{
Computes the euclidean distances for n pairs of points, e.g. pt0 vs pts[1,], pts[2,], pts[3,], etc. 
             It could be slow for large number of points.
}

