% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{dist.one0}
\alias{dist.one0}
\title{Computes the euclidean distances for n pairs of points using faster C codes}
\usage{
dist.one0(tgt, pts)
}
\arguments{
\item{tgt}{target, a vector of length p}

\item{pts}{n x p matrix, where each row is a point in p-dimensional space}
}
\value{
matrix of distance
}
\description{
Computes the euclidean distances for *all* pairs of points, e.g. pt0 vs pts[1,], pts[2,], pts[3,], etc.
             Similar to dist.one(). This version uses the build-in dist() function
}

