% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density_fun.r
\name{dmvnorm.hext}
\alias{dmvnorm.hext}
\title{Multivariable normal density for a hidden state evaluated at x}
\usage{
dmvnorm.hext(x, j, model, Log = FALSE)
}
\arguments{
\item{x}{numeric matrix}

\item{j}{hidden state index}

\item{model}{HMM model}

\item{Log}{boolean whether log is used, default to FALSE}
}
\value{
vector of proabilities, length equal the number of rows in x
}
\description{
Multivariable normal density for a hidden state evaluated at x
}
\details{
used in generic.density
}

