% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density_fun.r
\name{Dmvnorm}
\alias{Dmvnorm}
\title{Multivariable normal density, similar to the similar named function in mvtnorm}
\usage{
Dmvnorm(x, mean, sigma, log = TRUE, num.cutoff = 1e-100)
}
\arguments{
\item{x}{numeric matrix}

\item{mean}{mean vector}

\item{sigma}{covariance matrix}

\item{log}{boolean whether log is used, default to FALSE}

\item{num.cutoff}{numeric cutoff value}
}
\value{
vector of proabilities, length equal the number of rows in x
}
\description{
Multivariable normal density, similar to the similar named function in mvtnorm
}
\details{
use ginv from MASS to compute inverse matrix
}

