% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{Eln}
\alias{Eln}
\title{Compute the logarithm of a value}
\usage{
Eln(x)
}
\arguments{
\item{x}{numeric value}
}
\value{
log(x) if x > 0, otherwise it gives -Inf
}
\description{
Internal function to compute the log of a value
}

