% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{Elnsum}
\alias{Elnsum}
\title{Compute log(X+Y) given log(X) and log(Y)}
\usage{
Elnsum(lX, lY)
}
\arguments{
\item{lX}{log(X), a scalar}

\item{lY}{log(Y), a scalar}
}
\value{
numeric value
}
\description{
Calculate log(X+Y), Scalar version
}

