% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{Elnsum.vec}
\alias{Elnsum.vec}
\title{Compute log(X+Y) given log(X) and log(Y), vectorized version}
\usage{
Elnsum.vec(lx)
}
\arguments{
\item{lx}{vector of log(x1), log(x2), .... log(xn)}
}
\value{
numeric value
}
\description{
Calculate log(x1 + x2 + ... = xn), vectorized version
}

