% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{expected.neighbors}
\alias{expected.neighbors}
\title{Compute the expected number of neighbors in a ball of given radius}
\usage{
expected.neighbors(dat, ball.r)
}
\arguments{
\item{dat}{numeric matrix of size n x p, where n is the nubmer of points in p dimensional space}

\item{ball.r}{radius of a p-dimensional sphere}
}
\value{
number of points in the sphere
}
\description{
compute the expected number of neighbors
}

