% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Max_Sep.R
\name{F.potential}
\alias{F.potential}
\title{Field potential}
\usage{
F.potential(src, pt, q.i)
}
\arguments{
\item{src}{sources}

\item{pt}{point vector}

\item{q.i}{electric charges for each source point}
}
\value{
potential matrix
}
\description{
see my paper for detailed description
}

