% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_spec.R
\name{gen.init}
\alias{gen.init}
\title{Initialization of the HMM parameters}
\usage{
gen.init(p, K, D = 0, M = 0, A = NULL, Pi = NULL, Log = FALSE)
}
\arguments{
\item{p}{the entries of this vector give the dimension of the ith component of the obs}

\item{K}{number of hidden states}

\item{D}{vector of the size of the alphabets in each column of the multivariate discrete obs matrix}

\item{M}{number of gaussian mixtures}

\item{A}{transition probability matrix, if NULL, the init code is used to generate one}

\item{Pi}{initial state probabiilty, if NULL, the init code is used to generate one}

\item{Log}{boolean, whether log is used}
}
\value{
list consists of the initial state probability Pi, the transition matrix A, and emission probability b
}
\description{
Initialization of the HMM parameters
}
\details{
see code for details
}

