% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density_fun.r
\name{generic.density}
\alias{generic.density}
\title{Generic density function for HMM's emission probability}
\usage{
generic.density(x, j, model, NegInf = -1000)
}
\arguments{
\item{x}{numeric matrix, size T x p, where T is the number of observations, p the dimension of the obs data}

\item{j}{hidden state index}

\item{model}{current model as defined in hmmspec.gen(..)}
}
\value{
the emssion density of hidden state j
}
\description{
Generic density function for HMM's emission probability
}
\details{
see example code for usage
}

