% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Topcalls.r
\name{hextfit}
\alias{hextfit}
\title{Train HMM with model specificaiton provided in `model`}
\usage{
hextfit(x, model, lock.transition = FALSE, tol = 0.1, maxit = 1000,
  verbose = TRUE)
}
\arguments{
\item{x}{vector or matrix of data}

\item{model}{model specification created by hmmspec.gen(..)}

\item{lock.transition}{boolean, whether to fix the transition matrix}

\item{tol}{numeric tolerance for convergence evaluation}

\item{maxit}{maximum number of EM iterations}

\item{verbose}{boolean, whether to print diagnostics}
}
\value{
list with elements that describe the trained model
}
\description{
Train HMM with model specificaiton provided in `model`
}
\details{
See example
}

