% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_spec.R
\name{hmmspec.gen}
\alias{hmmspec.gen}
\title{Model specificaiton in one object for hextfit call}
\usage{
hmmspec.gen(linit, ltrans, parms.emission, dens.emission = generic.density,
  rand.emission = NULL, mstep = mstep.generic, mtype = mtype, K, M, D,
  Log = TRUE, Intv = FALSE, wdf = 1)
}
\arguments{
\item{linit}{initial probability vector of hidden states, in log domain}

\item{ltrans}{transition probability matrix, in log domain}

\item{parms.emission}{paramters for emission distribution}

\item{dens.emission}{density function for emission distribution}

\item{rand.emission}{function used to generate obs from emission distribution}

\item{mstep}{density function for the M-step}

\item{mtype}{integer code to indicate the type of emission distribution}

\item{K}{number of hidden states}

\item{M}{number of gaussian mixtures}

\item{D}{vector of the size of the alphabets in each column of the multivariate discrete obs matrix}

\item{Log}{boolean, whether log is used}

\item{Intv}{interal for guassian distribution}

\item{wdf}{degree of freedom for Wishart distribution}
}
\value{
list with all these elements for hexfit
}
\description{
Model specificaiton in one object for hextfit call
}
\details{
see code for details
}

