% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{make_hext}
\alias{make_hext}
\title{Make a Text-Based Hex Sticker}
\usage{
make_hext(
  text_1 = "",
  text_2 = "",
  text_3 = "",
  text_4 = "",
  align_1 = c("centre", "left", "right"),
  align_2 = c("centre", "left", "right"),
  align_3 = c("centre", "left", "right"),
  align_4 = c("centre", "left", "right"),
  ...,
  .cat = TRUE
)
}
\arguments{
\item{text_1, text_2, text_3, text_4}{Character. Text to display on each line
number.}

\item{align_1, align_2, align_3, align_4}{Character. Text alignment for each
line number. One of of \code{'centre'}, \code{'left'} or \code{'right'}. Defaults to
\code{'centre'}.}

\item{...}{Reserved for future use.}

\item{.cat}{Logical. Concatenate and print the four lines of the hex and
return the character vector silently? Otherwise return the character
vector. Defaults to \code{TRUE}.}
}
\description{
Make a Text-Based Hex Sticker
}
\details{
Lines 1 and 4 have a width of 8 characters. Lines 2 and 3 have width 10.
Your provided text cannot exceed this line width.
If you put text on line 4 then it will replace the border for that row.
}
\examples{
make_hext(
  "hello", "this", "is", "hext",
  "left", "right", "centre", "left"
)
}
