% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Max_Sep.R
\name{max.separate.N}
\alias{max.separate.N}
\title{Max Sep algorithm that find M points in the data cloud that are maximally separated}
\usage{
\method{max}{separate.N}(dat, M)
}
\arguments{
\item{dat}{data matrix}

\item{M}{number of maximally separate points}
}
\value{
matrix
}
\description{
see my paper for detailed description
}

