% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.r
\name{mstep.disc}
\alias{mstep.disc}
\title{M-step re-estimation function for univariate discrete emission distribution}
\usage{
mstep.disc(x, w, D)
}
\arguments{
\item{x}{numeric matrix, size sum(T) x 1. All sequences stacked onto one column}

\item{w}{weight matrix, size sum(T) x K, gamma(i, t)}

\item{D}{number of alphabets in x}

\item{model}{HMM model as defined in hmmspec.gen(..)}
}
\value{
emission slot of Hext object
}
\description{
Re-estimates the parameters of discrete emission distribution as part of the EM algorithm for HMM. No regularizer is used.
}
\author{
Horace W Tso
}

