% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.r
\name{mstep.gaussian.mix}
\alias{mstep.gaussian.mix}
\title{M-step re-estimation function for Gaussian Mixture emission distribution with inverse Wishart regularizer}
\usage{
mstep.gaussian.mix(x, w, model)
}
\arguments{
\item{x}{numeric matrix, size T x p, where T is the number of observations, p the dimension of the obs data}

\item{w}{weight matrix, gamma from EStep C function}

\item{model}{HMM model as defined in hmmspec.gen(..)}
}
\value{
emission slot of Hext object
}
\description{
Re-estimates the parameters of the gaussian mixture emission distribution as part of the EM algorithm for HMM.
}
\details{
Ref : Snoussi, Mohammad-Djafari (2005), Degeneracy and likelihood penalization
     in multivariate gaussian mixture models, preprint.  
     NOTE : gam.jk is *not* in log domain.
}
\author{
Horace W Tso
}

