% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.r
\name{mstep.generic}
\alias{mstep.generic}
\title{Generic M-step function to re-estimate the paramters of an emission probability distribution.}
\usage{
mstep.generic(x, wt, model)
}
\arguments{
\item{x}{numeric matrix, size T x p, where T is the number of observations, p the dimension of the obs data}

\item{wt}{weight matrix}

\item{model}{HMM model as defined in hmmspec.gen(..)}
}
\value{
the five types of M-step results
}
\description{
Re-estimate the parameters of an emission distribution as part of the EM algorithm for HMM. 
   This is called by the hext function.
}
\details{
see example code for usage
}
\author{
Horace W Tso
}

