% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.r
\name{mstep.mvbeta}
\alias{mstep.mvbeta}
\title{M-step re-estimation function for multivariate beta emission distribution}
\usage{
mstep.mvbeta(x, wt)
}
\arguments{
\item{x}{numeric matrix, size sum(T) x p}

\item{w}{weight matrix, size T x K}

\item{model}{HMM model as defined in hmmspec.gen(..)}
}
\value{
emission slot of Hext object
}
\description{
Re-estimates the parameters of multivariate beta emission distribution as part of the EM algorithm for HMM. No regularizer is used.
}
\details{
Needs the solnp function in the Rsolnp package.
}
\author{
Horace W Tso
}

