% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.r
\name{mstep.mvdisc}
\alias{mstep.mvdisc}
\title{M-step re-estimation function for multivariate discrete emission distribution}
\usage{
mstep.mvdisc(x, w, D)
}
\arguments{
\item{x}{numeric matrix, size sum(T) x p}

\item{w}{weight matrix, size T x K}

\item{D}{number of alphabets in x, a vector of length equal to ncol(x)}

\item{model}{HMM model as defined in hmmspec.gen(..)}
}
\value{
emission slot of Hext object
}
\description{
Re-estimates the parameters of discrete emission distribution as part of the EM algorithm for HMM. No regularizer is used.
}
\author{
Horace W Tso
}

