% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density_fun.r
\name{pmvnorm.intv}
\alias{pmvnorm.intv}
\title{Multivariable normal distribution at a point, calculate probability for a small interval around a point}
\usage{
pmvnorm.intv(x, mean, sigma, eps = 0.1)
}
\arguments{
\item{x}{numeric matrix}

\item{mean}{mean vector of the multivariate normal}

\item{sigma}{covariance matrix}

\item{eps}{small numeric neighborhood around x}
}
\value{
vector of proabilities, length equal the number of rows in x
}
\description{
Multivariable normal distribution at a point, calculate probability for a small interval around a point
}
\details{
used in generic.density
}

