% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxQ_func.R
\name{Qaux.beta}
\alias{Qaux.beta}
\title{Compute EM Auxiliary Q function for beta distribution}
\usage{
Qaux.beta(theta, x, wt)
}
\arguments{
\item{theta}{the shape parameters of beta distribution}

\item{x}{numeric vector to evaluate Aux Q function}

\item{wt}{weight vector}
}
\value{
numeric value of Aux Q function
}
\description{
See ref
}

