% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxQ_func.R
\name{Qaux.gamma}
\alias{Qaux.gamma}
\title{Compute EM Auxiliary Q function for gamma distribution}
\usage{
Qaux.gamma(theta, x, wt)
}
\arguments{
\item{theta}{the shape parameters of gamma distribution}

\item{x}{numeric vector to evaluate Aux Q function}

\item{wt}{weight vector}
}
\value{
numeric value of Aux Q function
}
\description{
See ref
}

