% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{shift.sequence}
\alias{shift.sequence}
\title{Internal function for testing the sensitivity of HMM parameters}
\usage{
shift.sequence(dat, TT, min.length = 10, shift = 5)
}
\arguments{
\item{dat}{: matrix with rows as observation and columns as features}

\item{TT}{: vector of sequence lengths, one entry per sequence}

\item{min.length}{: the minimum length of the *last* sequence}

\item{shift}{: the number of observations in the first sequence to discard so that the whole continuous observation is shifted forward}
}
\value{
list of x, N
}
\description{
Purpose of this function is to test how sensitive an hmm model is to 
different beginning of the sequence division. Its use is very limited
to the approach taken in DiscretRet.R. Some analysis needs to be made
when modeling long sequences that evaluates parameter stability when a 
few data points is added or taken away at the beginning and end of 
the sequence.
}

