% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simu.R
\name{sim.gaussmix}
\alias{sim.gaussmix}
\title{Simulate gaussian mixture data. Internal function called by simulate.gaussian.mix}
\usage{
sim.gaussmix(N, mu, sigma, cj)
}
\arguments{
\item{N}{number of data points to generate}

\item{mu}{a list of mean vectors, length equal to the number of mixtures}

\item{sigma}{a list of covariance matrices, length = no. of mixtures}

\item{cj}{a vector with the fraction to draw from each component, must sum to 1.0}
}
\value{
matrix of size N x p, where p is the dimension of the data
}
\description{
See example
}

