% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simu.R
\name{simulate.gauss.mix.matrix}
\alias{simulate.gauss.mix.matrix}
\title{Simulate gaussian mixture data, returns a matrix (wrapper function)}
\usage{
\method{simulate}{gauss.mix.matrix}(Len, mix.st, st.progression)
}
\arguments{
\item{Len}{vector of lengths of the sequences to generate}

\item{mix.st}{mixture structure M5 as defined in gen.init of the Hext package}

\item{st.progression}{integer vector for the state progression, e.g. {1, 3, 2, 1, 4}}

\item{SS}{vector of the length of each state, e.g. {100, 150, 130, 200, 180}}
}
\value{
matrix of size N x p, where p is the dimension of the data
}
\description{
wrapper function that calls simulate.gaussian.mix(..)
}
\details{
Length of TT must equal the length of st.progression.
}

