% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simu.R
\name{simulate.gaussian.mix}
\alias{simulate.gaussian.mix}
\title{Simulate gaussian mixture data (wrapper function)}
\usage{
\method{simulate}{gaussian.mix}(N, mix.st, state)
}
\arguments{
\item{N}{number of data points to generate}

\item{mix.st}{mixture structure M5 as defined in gen.init of the Hext package}

\item{state}{integer indicating one of the K states to use for this simulation}
}
\value{
matrix of size N x p, where p is the dimension of the data
}
\description{
wrapper function that calls sim.gaussmix(..)
}

