% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simu.R
\name{simulate.GMM.Nseq}
\alias{simulate.GMM.Nseq}
\title{Generate multiple sequences of multivariate gaussian mixture data}
\usage{
\method{simulate}{GMM.Nseq}(TT, LL, mix.st, st.progression)
}
\arguments{
\item{TT}{vector of the lengths of each sequence}

\item{LL}{vector of the lengths of each state in a sequence}

\item{mix.st}{mixture structure with mu, sigma, cj for each state and each mixture component}

\item{st.progression}{vector of integer indicating the state progression, e.g. {1, 3, 2, 1, 4}}
}
\value{
list with two elements : dat, the data matrix; ss : the state sequences
}
\description{
generate multiple sequences of multivariate gaussian mixture data. Calls simulate.gauss.mix.matrix()
}

