% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simu.R
\name{simulate.multinom}
\alias{simulate.multinom}
\title{Generate multinomial sequence data}
\usage{
\method{simulate}{multinom}(TT, D, st.progress)
}
\arguments{
\item{TT}{vector of the lengths of each state}

\item{D}{}

\item{st.progression}{vector of integer indicating the state progression, e.g. {1, 3, 2, 1, 4}}
}
\value{
matrix of size n x p
}
\description{
Note : TT and st.progress should have the same length
}

