% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simu.R
\name{simulate.mv.binary}
\alias{simulate.mv.binary}
\title{Generate multivariate binary data}
\usage{
\method{simulate}{mv.binary}(TT, p, K)
}
\arguments{
\item{TT}{vector of sequence lengths}

\item{p}{dimension of observation}

\item{K}{number of states}
}
\value{
matrix of size n x p
}
\description{
generate multivariate binary sequence
}

