% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{Transition.events.by.return}
\alias{Transition.events.by.return}
\title{Check each covariance matrix in the gaussian mixture specification}
\usage{
Transition.events.by.return(K, ss.vitb, dat)
}
\arguments{
\item{K}{number of viterbi states}

\item{ss.vitb}{vector of the viterbi-decoded states in the same time order as the observation sequence}

\item{dat}{data frame having forward return columns, whose names start with "F"}
}
\value{
list of Ret.by.state, Sd.by.state, event.counts, state.counts, enum
}
\description{
See example
}

