% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viterbi.R
\name{Viterbi.hext}
\alias{Viterbi.hext}
\title{Viterbi decoding algorithm}
\usage{
Viterbi.hext(model, newdata)
}
\arguments{
\item{model}{trained HMM model, output of hextfit}

\item{newdata}{data to decode with the trained model}
}
\value{
list with elements that describe the trained model
}
\description{
Viterbi decoding algorithm
}
\details{
the decoded sequence, likelihood, see code for details
}

