% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{weirdness}
\alias{weirdness}
\title{Inspect vector or matrix for number of NA, NaNs, and infinities}
\usage{
weirdness(x)
}
\arguments{
\item{x}{vector or matrix}
}
\value{
list of ina, number of NAs, inan, number of NaN, and iinf, number of infinities
}
\description{
Count the number of NAs, NaNs, and infinities in a matrix or vector
}

