% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hext-package.R
\docType{package}
\name{hext-package}
\alias{hext}
\alias{hext-package}
\title{hext: ASCII Hex Stickers}
\description{
Make a text-based hex logo, inspired by Yihui Xie and Charlie Gao.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray+hexlite@gmail.com}

}
\keyword{internal}
