% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hext-package.R
\docType{package}
\name{hext-package}
\alias{hext-package}
\title{hext: Make Text-Based Hex Logos}
\description{
Make a hex-sticker logo entirely from text. Inspired by Yihui Xie and Charlie Gao.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray+hext@gmail.com}

}
\keyword{internal}
